+++ src/scene/Light.h
@@ -72,7 +72,7 @@ extern EERIE_LIGHT * PDL[MAX_DYNLIGHTS];
 extern EERIE_LIGHT * GLight[MAX_LIGHTS];
 extern EERIE_LIGHT DynLight[MAX_DYNLIGHTS];
 extern long TOTPDL;
-extern long TOTIOPDL;
+extern size_t TOTIOPDL;
 
 ARX_HANDLE_TYPEDEF(long, LightHandle, -1)
